requirements.txt
jinja2
lightkube
lightkube-models
ops
pydantic
pytest-interface-tester
